Use functions instead of classes in FormState, fix type and value problems in FormFieldState, and move debounced to text inputs components
