+++ src/webapp/components/form/FieldWidget.tsx
@@ -1,16 +1,14 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 
-import i18n from "../../../utils/i18n";
 import { TextInput } from "../text-input/TextInput";
 import { MemberSelector } from "../member-selector/MemberSelector";
-import { FormField, FormFieldState } from "./FormFieldState";
-import { useDebounce } from "../../hooks/useDebounce";
 import { MultipleSelector } from "../selector/MultipleSelector";
 import { Selector } from "../selector/Selector";
 import { RadioButtonsGroup } from "../radio-buttons-group/RadioButtonsGroup";
 import { TextArea } from "../text-input/TextArea";
 import { DatePicker } from "../date-picker/DatePicker";
 import { Checkbox } from "../checkbox/Checkbox";
+import { FormFieldState, updateFieldState } from "./FormState";
 
 export type FieldWidgetProps = {
     onChange: (updatedField: FormFieldState) => void;
@@ -20,26 +18,16 @@ export type FieldWidgetProps = {
 
 export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
     const { field, onChange, disabled = false } = props;
-    const [textFieldValue, setTextFieldValue] = useState<string>(
-        field.type === "text" ? field.value : ""
-    );
-    const debouncedTextFieldValue = useDebounce(textFieldValue);
-
-    useEffect(() => {
-        if (debouncedTextFieldValue !== field.value && field.type === "text") {
-            onChange(FormField.updateState(field, debouncedTextFieldValue));
-        }
-    }, [debouncedTextFieldValue, field, onChange]);
 
     switch (field.type) {
         case "select": {
-            return field.multiple && field.options && Array.isArray(field.value) ? (
+            return field.multiple ? (
                 <MultipleSelector
                     id={field.id}
                     placeholder={field.placeholder}
                     label={field.label}
                     selected={field.value}
-                    onChange={newValue => onChange(FormField.updateState(field, newValue))}
+                    onChange={newValue => onChange(updateFieldState(field, newValue))}
                     options={field.options}
                     helperText={field.helperText}
                     errorText={field.errors ? field.errors.join("/n") : ""}
@@ -47,13 +35,13 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                     required={field.required && field.showIsRequired}
                     disabled={disabled}
                 />
-            ) : !field.multiple && field.options && typeof field.value === "string" ? (
+            ) : (
                 <Selector
                     id={field.id}
                     placeholder={field.placeholder}
                     label={field.label}
                     selected={field.value}
-                    onChange={newValue => onChange(FormField.updateState(field, newValue))}
+                    onChange={newValue => onChange(updateFieldState(field, newValue))}
                     options={field.options}
                     helperText={field.helperText}
                     errorText={field.errors ? field.errors.join("/n") : ""}
@@ -61,18 +49,16 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                     required={field.required && field.showIsRequired}
                     disabled={disabled}
                 />
-            ) : (
-                <span>{i18n.t("Error displaying this field in the form")}</span>
             );
         }
 
         case "radio": {
-            return field.options && typeof field.value === "string" ? (
+            return (
                 <RadioButtonsGroup
                     id={field.id}
                     label={field.label}
                     selected={field.value}
-                    onChange={event => onChange(FormField.updateState(field, event.target.value))}
+                    onChange={event => onChange(updateFieldState(field, event.target.value))}
                     options={field.options}
                     helperText={field.helperText}
                     errorText={field.errors ? field.errors.join("/n") : ""}
@@ -80,8 +66,6 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                     required={field.required && field.showIsRequired}
                     disabled={disabled}
                 />
-            ) : (
-                <span>{i18n.t("Error displaying this field in the form")}</span>
             );
         }
 
@@ -90,8 +74,8 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                 <TextArea
                     id={field.id}
                     label={field.label}
-                    value={textFieldValue}
-                    onChange={newValue => setTextFieldValue(newValue)}
+                    value={field.value}
+                    onChange={newValue => onChange(updateFieldState(field, newValue))}
                     helperText={field.helperText}
                     errorText={field.errors ? field.errors.join("/n") : ""}
                     error={field.errors && field.errors.length > 0}
@@ -102,8 +86,8 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                 <TextInput
                     id={field.id}
                     label={field.label}
-                    value={textFieldValue}
-                    onChange={event => setTextFieldValue(event.target.value)}
+                    value={field.value}
+                    onChange={newValue => onChange(updateFieldState(field, newValue))}
                     helperText={field.helperText}
                     errorText={field.errors ? field.errors.join("/n") : ""}
                     error={field.errors && field.errors.length > 0}
@@ -118,7 +102,7 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                     id={field.id}
                     label={field.label}
                     value={field.value}
-                    onChange={newValue => onChange(FormField.updateState(field, newValue))}
+                    onChange={newValue => onChange(updateFieldState(field, newValue))}
                     helperText={field.helperText}
                     errorText={field.errors ? field.errors.join("/n") : ""}
                     error={field.errors && field.errors.length > 0}
@@ -133,7 +117,7 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                     id={field.id}
                     label={field.label}
                     checked={field.value}
-                    onChange={newValue => onChange(FormField.updateState(field, newValue))}
+                    onChange={newValue => onChange(updateFieldState(field, newValue))}
                     helperText={field.helperText}
                     disabled={disabled}
                 />
@@ -147,7 +131,7 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                     placeholder={field.placeholder}
                     label={field.label}
                     selected={field.value}
-                    onChange={newValue => onChange(FormField.updateState(field, newValue))}
+                    onChange={newValue => onChange(updateFieldState(field, newValue))}
                     options={field.options}
                     helperText={field.helperText}
                     errorText={field.errors ? field.errors.join("/n") : ""}
@@ -157,8 +141,5 @@ export const FieldWidget: React.FC<FieldWidgetProps> = React.memo(props => {
                 />
             );
         }
-
-        default:
-            return <span>{i18n.t("Error displaying this field in the form")}</span>;
     }
 });
