+++ src/webapp/components/form/Form.tsx
@@ -1,10 +1,9 @@
 import React from "react";
 
 import { useForm } from "./useForm";
-import { FormState } from "./FormState";
+import { FormState, FormFieldState } from "./FormState";
 import { FormLayout } from "./FormLayout";
 import { FormSection } from "./FormSection";
-import { FormFieldState } from "./FormFieldState";
 import { Layout } from "../layout/Layout";
 
 export type FormProps = {
