+++ src/webapp/components/form/useForm.ts
@@ -1,15 +1,16 @@
 import { useCallback, useEffect, useState } from "react";
 
-import { Form, FormState } from "./FormState";
-import { FormFieldState } from "./FormFieldState";
+import { updateFormState, FormState, FormFieldState } from "./FormState";
+
+type State = {
+    formLocalState: FormState;
+    handleUpdateFormField: (updatedField: FormFieldState) => void;
+};
 
 export function useForm(
     formState: FormState,
     onFormChange: (newFormState: FormState, updatedField: FormFieldState) => void
-): {
-    formLocalState: FormState;
-    handleUpdateFormField: (updatedField: FormFieldState) => void;
-} {
+): State {
     const [formLocalState, setFormLocalState] = useState<FormState>(formState);
 
     useEffect(() => {
@@ -20,7 +21,7 @@ export function useForm(
 
     const handleUpdateFormField = useCallback(
         (updatedField: FormFieldState) => {
-            const newFormState = Form.updateFormState(formState, updatedField);
+            const newFormState = updateFormState(formState, updatedField);
             setFormLocalState(newFormState);
             onFormChange(newFormState, updatedField);
         },
